Refactor: moved some more login and logout login into user controller
